AzureFunctions/api/uploadImage/requirements.txt
azure-core==1.32.0
azure-functions==1.21.3
azure-storage-blob==12.24.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
Flask==3.1.0
frozendict==2.4.6
git-filter-repo==2.45.0
html5lib==1.1
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==3.0.2
multitasking==0.0.11
numpy==2.1.2
pandas==2.2.3
peewee==3.17.7
pillow==11.0.0
platformdirs==4.3.6
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
ta==0.11.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
Werkzeug==3.1.3
yfinance==0.2.48
